requirements.txt
# Core dependencies
playwright>=1.40.0
crawl4ai>=0.3.0
pydantic>=2.0.0
structlog>=23.1.0
python-dotenv>=1.0.0
aiofiles>=23.1.0
pytz>=2023.3  # Para timezone brasileiro nos logs

# HTTP/API
aiohttp>=3.8.0
httpx>=0.24.0
requests>=2.31.0

# FastAPI dependencies for API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Supabase dependencies
supabase>=2.0.0
postgrest>=0.13.0

# Email handling
imaplib2>=0.57
email-validator>=2.0.0

# Data processing
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
